+++ apps/synapse-portal-framework/src/components/adknowledge/AdknowledgeHeader/AdknowledgeHeader.tsx
@@ -3,6 +3,8 @@ import { Box, Stack, Typography } from '@mui/material'
 import HeaderSearchBox from '@/components/HeaderSearchBox'
 import { TypeAnimation } from 'react-type-animation'
 import styles from './AdknowledgeHeader.module.scss'
+import { WordPressLatestPostChip } from 'synapse-react-client/components/WordPress/WordPressLatestPostChip'
+
 const AdknowledgeHeader = (): React.ReactNode => {
   const searchPlaceholder = 'Search for...'
 
@@ -45,6 +47,7 @@ const AdknowledgeHeader = (): React.ReactNode => {
   return (
     <header className={styles.adknowledgeHeader}>
       <Stack className={styles.adknowledgeHeaderStack}>
+        <WordPressLatestPostChip wordpressSiteUrl="https://news.adknowledgeportal.org" />
         <Box className={styles.adknowledgeHeaderContentBox}>{content}</Box>
         <HeaderSearchBox
           searchExampleTerms={searchExampleTerms}
