+++ packages/synapse-react-client/src/components/SynapseHomepageV2/WordPressNews.tsx
@@ -1,10 +1,7 @@
-import { useQuery } from '@tanstack/react-query'
 import { Box } from '@mui/material'
-import {
-  WordPressCategory,
-  WordPressNewsItem,
-  WordPressPost,
-} from './WordPressNewsItem'
+import { useWordPressCategories } from '../WordPress/useWordPressCategories'
+import { useWordPressPosts } from '../WordPress/useWordPressPosts'
+import { WordPressNewsItem } from './WordPressNewsItem'
 import styles from './HotdropsAndNews.module.scss'
 
 export type WordPressNewsProps = {
@@ -16,51 +13,14 @@ export type WordPressNewsProps = {
   variant?: 'adkp'
 }
 
-async function fetchWordPressPosts(
-  wordpressSiteUrl: string,
-  postCount: number,
-): Promise<WordPressPost[]> {
-  const url = new URL(`${wordpressSiteUrl}/wp-json/wp/v2/posts`)
-  url.searchParams.set('per_page', String(postCount))
-  url.searchParams.set('orderby', 'date')
-  url.searchParams.set('order', 'desc')
-  url.searchParams.set('_fields', 'id,date,link,title,categories')
-  const response = await fetch(url.toString())
-  if (!response.ok) {
-    throw new Error(`Failed to fetch WordPress posts: ${response.statusText}`)
-  }
-  return response.json() as Promise<WordPressPost[]>
-}
-
-async function fetchWordPressCategories(
-  wordpressSiteUrl: string,
-): Promise<WordPressCategory[]> {
-  const url = new URL(`${wordpressSiteUrl}/wp-json/wp/v2/categories`)
-  url.searchParams.set('_fields', 'id,name,slug')
-  const response = await fetch(url.toString())
-  if (!response.ok) {
-    throw new Error(
-      `Failed to fetch WordPress categories: ${response.statusText}`,
-    )
-  }
-  return response.json() as Promise<WordPressCategory[]>
-}
-
 export function WordPressNews({
   wordpressSiteUrl,
   postCount = 10,
   showCategoryChips = true,
   variant,
 }: WordPressNewsProps) {
-  const { data: posts } = useQuery({
-    queryKey: ['wordpressPosts', wordpressSiteUrl, postCount],
-    queryFn: () => fetchWordPressPosts(wordpressSiteUrl, postCount),
-  })
-
-  const { data: categories } = useQuery({
-    queryKey: ['wordpressCategories', wordpressSiteUrl],
-    queryFn: () => fetchWordPressCategories(wordpressSiteUrl),
-  })
+  const { data: posts } = useWordPressPosts(wordpressSiteUrl, postCount)
+  const { data: categories } = useWordPressCategories(wordpressSiteUrl)
 
   const categoryMap = new Map<number, string>(
     (categories ?? []).map(c => [c.id, c.name]),
